improved tree positioner api
